requirements.txt
pytest==7.4.3
Appium-Python-Client==3.1.1
selenium==4.35.0
pytest-html==4.1.1
pytest-xdist==3.3.1
